scripts: improve clean script usability

+++ scripts/clean_all.bash
@@ -3,10 +3,10 @@ set -e
 
 # parse help argument
 if [[ $1 == "-h" ]] || [[ $1 == "--help" ]]; then
-  echo -e "Usage: clean_all.bash [option...] \t This script cleans the build for both workspaces" >&2
+  echo -e "Usage: clean_all.bash [option...] \t This script cleans the build of the workspace(s)" >&2
   echo
   echo -e "\t--delete_ros1_bridge \t Deletes the ros1_bridge repository under the ROS2 workspace"
-  echo -e "\t--ros1_ws_dir \t\t Location of the ROS(1) workspace where one has cloned px4_ros_com 'ros1' branch. Default: $HOME/px4_ros_com_ros1"
+  echo -e "\t--delete_ros1_ws_build \t\t Deletes 'build', 'devel', 'install' and 'logs' folder on the given location of the ROS(1) workspace where one has cloned px4_ros_com 'ros1' branch. If none is set, defaults to $HOME/px4_ros_com_ros1"
   echo
   exit 0
 fi
@@ -30,21 +30,21 @@ ROS2_REPO_DIR=$(cd "$(dirname "$SCRIPT_DIR")" && pwd)
 ROS2_WS_SRC_DIR=$(cd "$(dirname "$ROS2_REPO_DIR")" && pwd)
 ROS2_WS_DIR=$(cd "$(dirname "$ROS2_WS_SRC_DIR")" && pwd)
 
-# ROS1 dirs (one can pass the ROS1 workspace dir using '--ros1_ws_dir <ws_dir>')
-ROS1_WS_DIR=${ros1_ws_dir:-"$(cd "$HOME/px4_ros_com_ros1" && pwd)"}
-ROS1_REPO_DIR=$(cd $ROS1_WS_DIR/src/px4_ros_com && pwd)
+# ROS (1) workspace dir to delete generated build files (one can pass the workspace dir using '--delete_ros1_ws_build <ws_dir>')
+ROS1_WS_DIR=${delete_ros1_ws_build:-""}
 
+echo -e "\t - ROS2:"
 # clean micrortps_agent generated source code
 if [ -d $ROS2_WS_SRC_DIR/px4_ros_com/src/micrortps_agent ]; then
   cd $ROS2_WS_SRC_DIR/px4_ros_com/src && sudo rm -rf micrortps_agent
-  echo -e "\t - Deleted $ROS2_WS_SRC_DIR/px4_ros_com/src/micrortps_agent"
+  echo -e "\t -- Deleted $ROS2_WS_SRC_DIR/px4_ros_com/src/micrortps_agent"
   CLEAN_HISTORY=1
 fi
 
 # if '--delete_ros1_bridge' set, delete the ros1_bridge repo
-if [ -v $delete_ros1_bridge ] && [ -d $ROS2_WS_SRC_DIR/ros1_bridge ]; then
+if [ ! -v $delete_ros1_bridge ] && [ -d $ROS2_WS_SRC_DIR/ros1_bridge ]; then
   cd $ROS2_WS_SRC_DIR && sudo rm -rf ros1_bridge
-  echo -e "\t - Deleted $ROS2_WS_SRC_DIR/ros1_bridge"
+  echo -e "\t -- Deleted $ROS2_WS_SRC_DIR/ros1_bridge"
   CLEAN_HISTORY=1
 fi
 
@@ -53,20 +53,23 @@ dirs=("$ROS2_WS_DIR/build" "$ROS2_WS_DIR/install" "$ROS2_WS_DIR/log")
 for dir in "${dirs[@]}"; do
   if [ -d $dir ]; then
     sudo rm -rf $dir
-    echo -e "\t - Deleted $dir"
+    echo -e "\t -- Deleted $dir"
     CLEAN_HISTORY=1
   fi
 done
 
-# delete build, install and log folders in the ROS1 workspace
-dirs=("$ROS1_WS_DIR/build" "$ROS1_WS_DIR/install" "$ROS1_WS_DIR/log")
-for dir in "${dirs[@]}"; do
-  if [ -d $dir ]; then
-    sudo rm -rf $dir
-    echo -e "\t - Deleted $dir"
-    CLEAN_HISTORY=1
-  fi
-done
+# delete build, devel (if built with catkin), install and log folders in the ROS (1) workspace
+echo -e "\t - ROS(1):"
+if [ ! -v $ROS1_WS_DIR ]; then
+  dirs=("$ROS1_WS_DIR/build" "$ROS1_WS_DIR/devel" "$ROS1_WS_DIR/install" "$ROS1_WS_DIR/log")
+  for dir in "${dirs[@]}"; do
+    if [ -d $dir ]; then
+      sudo rm -rf $dir
+      echo -e "\t -- Deleted $dir"
+      CLEAN_HISTORY=1
+    fi
+  done
+fi
 
 if [ $CLEAN_HISTORY = 1 ]; then
   echo "Done cleaning build!..."
